Move reviewer's guide tests into own module
